requirements.txt
streamlit
fastapi
uvicorn
pandas
plotly
numpy
scikit-learn
openai
python-dotenv
langchain
langchain-experimental
langchain-openai
tabulate